hw07: file copying
