+++ hw07_file_copying/go.mod
@@ -2,4 +2,7 @@ module github.com/alexMolokov/hw-go-otus/hw07_file_copying
 
 go 1.16
 
-require github.com/cheggaaa/pb/v3 v3.0.8
+require (
+	github.com/cheggaaa/pb/v3 v3.0.8
+	github.com/stretchr/testify v1.7.0
+)
